box models: reorganization for context objects
allows to better decouple the problem from the model and will also
make the generic implementation of multiple spatial and temporal
discretizations _much_ easier.

git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@6671 2fb0f335-1f38-0410-981e-8018bf24f1b0

+++ dumux/common/timemanager.hh
@@ -379,7 +379,7 @@ public:
             if (verbose_) {
                 std::cout <<
                     boost::format("Time step %d done. Wall time:%.4g, time:%.4g, time step size:%.4g\n")
-                    %timeStepIndex()%timer.elapsed()%time()%dt;
+                    %timeStepIndex()%double(timer.elapsed())%double(time())%double(dt);
             }
 
             // write restart file if mandated by the problem
